Create testfile in no workspaces and copy paste
